feat: implement hash, compress/decompress & split code into modules
